TRIPLEX-820 Refactor Link component stories: restructure and enhance documentation with new examples, update visual test screenshots, and improve glob handling for testing specific components.
