+++ .gitignore
@@ -48,3 +48,6 @@ build-storybook.log
 
 # Visual regression testing (auto-generated, not committed)
 __screenshots__/__diff__/
+
+# Claude Code worktrees (temporary isolated repo copies used by agents)
+.claude/worktrees/
