+++ .storybook/main.ts
@@ -9,7 +9,9 @@ const config: StorybookConfig = {
         name: "@storybook/react-vite",
         options: {},
     },
-    stories: ["../stories/**/*.stories.@(ts|tsx|mdx)", "../stories/**/*.mdx"],
+    // stories: ["../stories/**/*.stories.@(ts|tsx|mdx)", "../stories/**/*.mdx"],
+    // для отладки конкретного компонента
+    stories: ["../stories/Link/Link.stories.tsx"],
     typescript: {
         reactDocgen: "react-docgen-typescript",
         reactDocgenTypescriptOptions: {
